server/requirements.txt
RPi.GPIO
websockets
